# db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package db

import (
	"time"

	"github.com/google/uuid"
	null "github.com/guregu/null/v5"
)

type Article struct {
	Uuid            uuid.UUID   `json:"uuid"`
	Name            string      `json:"name"`
	Desc            null.String `json:"desc"`
	PurchasePrice   float64     `json:"purchase_price"`
	ResellPrice     float64     `json:"resell_price"`
	ArticleTypeUuid uuid.UUID   `json:"article_type_uuid"`
}

type ArticleTransaction struct {
	Uuid            uuid.UUID `json:"uuid"`
	ArticleUuid     uuid.UUID `json:"article_uuid"`
	TransactionUuid uuid.UUID `json:"transaction_uuid"`
	Amount          int32     `json:"amount"`
	Price           float64   `json:"price"`
}

type ArticleType struct {
	Uuid          uuid.UUID   `json:"uuid"`
	Name          string      `json:"name"`
	Desc          null.String `json:"desc"`
	IconCodepoint int32       `json:"icon_codepoint"`
	Color         string      `json:"color"`
}

type Event struct {
	Uuid     uuid.UUID   `json:"uuid"`
	Name     string      `json:"name"`
	Desc     null.String `json:"desc"`
	FromDate time.Time   `json:"from_date"`
	ToDate   time.Time   `json:"to_date"`
}

type Transaction struct {
	Uuid  uuid.UUID `json:"uuid"`
	Date  time.Time `json:"date"`
	Price float64   `json:"price"`
}