# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend vor Rupay",
        "title": "Rupay Backend",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8888",
    "basePath": "/api",
    "paths": {
        "/article": {
            "post": {
                "description": "Create a new article with the provided payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Create a new article",
                "parameters": [
                    {
                        "description": "Create article payload",
                        "name": "article",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.CreateArticle"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created article",
                        "schema": {
                            "$ref": "#/definitions/db.Article"
                        }
                    },
                    "400": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to create article",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/articles": {
            "get": {
                "description": "Get a list of all articles",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Retrieve all articles",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved all articles",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Article"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve articles",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/articles/{articleId}": {
            "get": {
                "description": "Retrieve an article using its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Get an article by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article ID",
                        "name": "articleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved article",
                        "schema": {
                            "$ref": "#/definitions/db.Article"
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve article",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an article with the provided ID and payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Update an existing article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article ID",
                        "name": "articleId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update article payload",
                        "name": "article",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.UpdateArticle"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated article",
                        "schema": {
                            "$ref": "#/definitions/db.Article"
                        }
                    },
                    "400": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update article",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove an article using its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Delete an article by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Article ID",
                        "name": "articleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted article"
                    },
                    "404": {
                        "description": "Article not found",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete article",
                        "schema": {
                            "$ref": "#/definitions/e.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "db.Article": {
            "type": "object",
            "properties": {
                "article_type_uuid": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "purchase_price": {
                    "type": "number"
                },
                "resell_price": {
                    "type": "number"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "e.ErrorDetail": {
            "type": "object",
            "properties": {
                "field": {
                    "description": "Field causing the error",
                    "type": "string"
                },
                "issue": {
                    "description": "Description of the issue with the field",
                    "type": "string"
                }
            }
        },
        "e.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code",
                    "type": "string"
                },
                "details": {
                    "description": "List of field-specific errors",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/e.ErrorDetail"
                    }
                },
                "error": {
                    "description": "Stack Trace",
                    "type": "string"
                },
                "message": {
                    "description": "Human-readable error message",
                    "type": "string"
                }
            }
        },
        "schemas.CreateArticle": {
            "type": "object",
            "required": [
                "article_type_uuid",
                "name",
                "purchase_price",
                "resell_price"
            ],
            "properties": {
                "article_type_uuid": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "purchase_price": {
                    "type": "number"
                },
                "resell_price": {
                    "type": "number"
                }
            }
        },
...
        "schemas.UpdateArticle": {
            "type": "object",
            "properties": {
                "article_type_uuid": {
                    "$ref": "#/definitions/uuid.NullUUID"
                },
                "desc": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "purchase_price": {
                    "type": "number"
                },
                "resell_price": {
                    "type": "number"
                }
            }
        },
...
        "uuid.NullUUID": {
            "type": "object",
            "properties": {
                "uuid": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if UUID is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}